+++ backend/drizzle/production/0002_nappy_bishop.sql
@@ -1,0 +1,1 @@
+ALTER TYPE "main"."files-types" ADD VALUE 'round_instructions' BEFORE 'prover_prompt';